requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
SQLAlchemy>=2.0
psycopg2-binary
python-multipart
